+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- Increase default `max_buffer` for `stdin`, `file` and `http_client` inputs.
+
 ## 0.13.3 - 2018-06-06
 
 ### Added